OBD2_PIDs.h kann jetzt in Xcode und Arduino benutzt werden.

+++ CarHud/CHOBD2.swift
@@ -63,7 +63,7 @@ import Foundation
 
 
 struct OBDValue {
-    var pid: UInt8
+    var pid: Int32
     var size: UInt8
 }
